+++ src/_utils/icon/index.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 import BaseIcon from './BaseIcon'
 
 const StyledBaseIcon = styled(BaseIcon)`
-  & {
+  &.kirk-icon-wrapper {
     display: inline-block;
     position: relative;
   }