+++ src/typings/icons.d.ts
@@ -1,6 +1,7 @@
 declare interface Icon {
   readonly size?: number | string
   readonly className?: Classcat.Class
+  readonly iconClassName? : Classcat.Class
   readonly title?: string
   readonly iconColor?: string
   readonly badgeAriaLabel?: string